# auth/src/swagger/swagger.yml
openapi: 3.0.0
info:
  title: Software2 API Documentation
  description: |
    This is the API documentation for the Software2 project. It provides details about the available authentication, permission, and role endpoints, their usage, and expected inputs/outputs.
  termsOfService: https://example.com/terms/
  contact:
    email: support@software2.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.software2.com
    description: Production server
...
paths:
...
  /users:
    get:
      tags:
        - user
      summary: Get all users
      description: Retrieve a list of all users in the system.
      operationId: getAllUsers
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Server error
    post:
      tags:
        - user
      summary: Create a new user
      description: Add a new user to the system.
      operationId: createUser
      requestBody:
        description: User object to be created
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  example: "john.doe@example.com"
                current_password:
                  type: string
                  example: "password123"
                number:
                  type: string
                  example: "+573001234567"
                rolId:
                  type: string
                  example: "6813d88bd0cb1281aaa26120"
                status:
                  type: string
                  example: "PENDING"
      responses:
        '201':
          description: User created successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
  /users/{id}:
    get:
      tags:
        - user
      summary: Get user by ID
      description: Retrieve a specific user by their ID.
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the user
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
        '500':
          description: Server error
    put:
      tags:
        - user
      summary: Update a user
      description: Update the details of an existing user.
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the user
      requestBody:
        description: User object to be updated
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                  example: "Jane Doe"
                email:
                  type: string
                  example: "jane.doe@example.com"
                current_password:
                  type: string
                  example: "newpassword123"
                number:
                  type: string
                  example: "+573001234568"
                rolId:
                  type: string
                  example: "6813d88bd0cb1281aaa26120"
                status:
                  type: string
                  example: "ACTIVE"
      responses:
        '200':
          description: User updated successfully
        '400':
          description: Invalid input
        '500':
          description: Server error
    delete:
      tags:
        - user
      summary: Delete a user
      description: Remove a user from the system.
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the user
      responses:
        '200':
          description: User deleted successfully
        '404':
          description: User not found
        '500':
          description: Server error
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "12345"
        fullname:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "john.doe@example.com"
        number:
          type: string
          example: "+573001234567"
        rolId:
          type: string
          example: "6813d88bd0cb1281aaa26120"
        status:
          type: string
          example: "ACTIVE"